+++ rest_framework/__init__.py
@@ -1,4 +1,4 @@
-__version__ = '2.3.3'
+__version__ = '2.3.4'
 
 VERSION = __version__  # synonym